+++ CHANGELOG.md
@@ -4,8 +4,9 @@
 
 ### Added
 
-- Use vault to generate secrets for minio using kv
+- Use vault to generate secrets for minio using kv #16
 - Code to support successful execution of nomad mc job and tests when consul_acl_default_policy is deny #26
+- Updated README.md #21
 
 ## [0.0.3]